quieted warning about format string type inconsistency

+++ crawl-ref/source/rltiles/tool/tile_list_processor.cc
@@ -528,8 +528,8 @@ bool tile_list_processor::process_line(char *read_line, const char *list_file,
                     if (val < 0 || val > 255)
                     {
                         fprintf(stderr,
-                                "Error (%s:%d): Arg %d must be 0-255.\n",
-                                list_file, line, i);
+                                "Error (%s:%d): Arg %lu must be 0-255.\n",
+                                list_file, line, (unsigned long)i);
                     }
 
                     cols[col_idx][comp_idx++] = static_cast<unsigned char>(val);
